fix(integrations): robust OAuth token handling and shared crypto utility
- Centralized OAuth encryption/decryption logic into `server/utils/oauthCrypto.ts`.
- Updated `decryptToken` to safely handle plaintext tokens (no colons) and malformed tokens (validation of hex parts) with descriptive warnings.
- Enforced 32-byte key length validation for `OAUTH_ENCRYPTION_KEY`.
- Refactored Google Calendar and Google Photos integrations to use the shared utility.
- Fixed 500 error during initial Google integration setup flow.
