+++ server/integrations/google-calendar/oauth.ts
@@ -2,83 +2,10 @@ import type { OAuth2Client } from "google-auth-library";
 
 import { consola } from "consola";
 import { google } from "googleapis";
-import { Buffer } from "node:buffer";
-import crypto from "node:crypto";
 
 import type { TokenInfo } from "./types";
 
-/**
- * Encrypt OAuth token using AES-256-GCM
- * @param token - Plain text token to encrypt
- * @returns Encrypted token in format: iv:authTag:encrypted
- */
-export function encryptToken(token: string): string {
-  const algorithm = "aes-256-gcm";
-  const key = getEncryptionKey();
-  const iv = crypto.randomBytes(16);
-
-  const cipher = crypto.createCipheriv(algorithm, key, iv);
-
-  let encrypted = cipher.update(token, "utf8", "hex");
-  encrypted += cipher.final("hex");
-  const authTag = cipher.getAuthTag();
-
-  return `${iv.toString("hex")}:${authTag.toString("hex")}:${encrypted}`;
-}
-
-/**
- * Decrypt OAuth token using AES-256-GCM
- * @param encryptedToken - Encrypted token in format: iv:authTag:encrypted
- * @returns Decrypted plain text token
- */
-export function decryptToken(encryptedToken: string): string {
-  // If token doesn't look like it's encrypted (no colons), return as is
-  // This happens during initial setup when frontend passes plaintext tokens
-  if (!encryptedToken.includes(":")) {
-    return encryptedToken;
-  }
-
-  const algorithm = "aes-256-gcm";
-  const key = getEncryptionKey();
-  const [ivHex, authTagHex, encrypted] = encryptedToken.split(":");
-
-  if (!ivHex || !authTagHex || !encrypted) {
-    // If it has colons but not the right number of parts, it's still probably not a valid encrypted token
-    // but might be a malformed plaintext token. We'll return as is to be safe during setup.
-    return encryptedToken;
-  }
-
-  const iv = Buffer.from(ivHex, "hex");
-  const authTag = Buffer.from(authTagHex, "hex");
-  const decipher = crypto.createDecipheriv(algorithm, key, iv);
-  decipher.setAuthTag(authTag);
-
-  let decrypted = decipher.update(encrypted, "hex", "utf8");
-  decrypted += decipher.final("utf8");
-
-  return decrypted;
-}
-
-/**
- * Get encryption key from environment, generating one if not set (dev only)
- * @returns 32-byte encryption key buffer
- */
-function getEncryptionKey(): Buffer {
-  const config = useRuntimeConfig();
-  const envKey = config.oauthEncryptionKey as string;
-
-  if (envKey) {
-    return Buffer.from(envKey, "hex");
-  }
-
-  // Development fallback - generate temporary key
-  if (import.meta.dev) {
-    consola.warn("OAUTH_ENCRYPTION_KEY not set, using temporary key (development only)");
-    return crypto.randomBytes(32);
-  }
-
-  throw new Error("OAUTH_ENCRYPTION_KEY environment variable is required");
-}
+export { decryptToken, encryptToken, getEncryptionKey } from "../../utils/oauthCrypto";
 
 /**
  * Create Google OAuth2 client with credentials from environment
